upgrade Gradle to 6.7.1 upgrade SpringBoot to 2.4.0 use Map for storing data
