+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'price-h2'
+rootProject.name = 'price-list'
