+++ src/main/java/vlaship/price/entity/User.java
@@ -1,0 +1,16 @@
+package vlaship.price.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class User {
+    private String name;
+    private String password;
+}
+
