+++ src/main/java/vlaship/price/exception/DownloadException.java
@@ -1,0 +1,7 @@
+package vlaship.price.exception;
+
+public class DownloadException extends RuntimeException {
+    public DownloadException(String message) {
+        super(message);
+    }
+}
