--- src/main/java/vlaship/price/h2/App.java
@@ -1,13 +1,0 @@
-package vlaship.price.h2;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class App {
-
-	public static void main(String[] args) {
-		SpringApplication.run(App.class, args);
-	}
-
-}
