--- src/main/java/vlaship/price/h2/entity/Role.java
@@ -1,6 +1,0 @@
-package vlaship.price.h2.entity;
-
-public enum Role {
-    VIEW,
-    ADD;
-}
