--- src/main/java/vlaship/price/h2/entity/User.java
@@ -1,25 +1,0 @@
-package vlaship.price.h2.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.bson.types.ObjectId;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Document
-public class User {
-
-    @Id
-    private ObjectId id;
-    private String name;
-    private String password;
-    private Role role;
-
-}
-
