--- src/main/java/vlaship/price/h2/exception/DownloadException.java
@@ -1,7 +1,0 @@
-package vlaship.price.h2.exception;
-
-public class DownloadException extends RuntimeException {
-    public DownloadException(String message) {
-        super(message);
-    }
-}
