--- src/main/java/vlaship/price/h2/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package vlaship.price.h2.repository;
-
-import org.bson.types.ObjectId;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import vlaship.price.h2.entity.User;
-
-import java.util.Optional;
-
-public interface UserRepository extends MongoRepository<User, ObjectId> {
-    Optional<User> findByName(String name);
-}
