--- src/main/java/vlaship/price/h2/security/SecurityConfigProd.java
@@ -1,57 +1,0 @@
-package vlaship.price.h2.security;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Profile("!dev")
-@Configuration
-@EnableWebSecurity
-public class SecurityConfigProd extends WebSecurityConfigurerAdapter {
-
-    private final SimpleDbUserDetailsService userDetailsService;
-
-    @Override
-    public void configure(final HttpSecurity http) throws Exception {
-        http
-                .csrf().disable()
-                .headers().frameOptions().disable() // h2 console
-                .and()
-                .authorizeRequests()
-                .antMatchers("/h2/**").permitAll()
-                .anyRequest().authenticated()
-                .and()
-                .formLogin();
-    }
-
-    @Override
-    protected void configure(final AuthenticationManagerBuilder auth) {
-        auth.authenticationProvider(authenticationProvider());
-    }
-
-    @Bean
-    public DaoAuthenticationProvider authenticationProvider() {
-        DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
-        authProvider.setUserDetailsService(userDetailsService);
-        authProvider.setPasswordEncoder(encoder());
-        return authProvider;
-    }
-
-    @Bean
-    public PasswordEncoder encoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Autowired
-    public SecurityConfigProd(final SimpleDbUserDetailsService userDetailsService) {
-        this.userDetailsService = userDetailsService;
-    }
-}
