--- src/main/java/vlaship/price/h2/service/DownloadCsvService.java
@@ -1,7 +1,0 @@
-package vlaship.price.h2.service;
-
-import java.util.List;
-
-public interface DownloadCsvService {
-    List<String> download();
-}
