--- src/main/java/vlaship/price/h2/service/ParserService.java
@@ -1,9 +1,0 @@
-package vlaship.price.h2.service;
-
-import vlaship.price.h2.entity.Product;
-
-import java.util.List;
-
-public interface ParserService {
-    List<Product> parse(List<String> lines);
-}
