--- src/main/java/vlaship/price/h2/service/UpdateDbService.java
@@ -1,5 +1,0 @@
-package vlaship.price.h2.service;
-
-public interface UpdateDbService {
-    void update();
-}
