--- src/main/java/vlaship/price/h2/service/UploadDataService.java
@@ -1,9 +1,0 @@
-package vlaship.price.h2.service;
-
-import vlaship.price.h2.entity.Product;
-
-import java.util.List;
-
-public interface UploadDataService {
-    void upload(List<Product> products);
-}
