+++ src/main/java/vlaship/price/service/DownloadCsvService.java
@@ -1,0 +1,7 @@
+package vlaship.price.service;
+
+import java.util.List;
+
+public interface DownloadCsvService {
+    List<String> download();
+}
