+++ src/main/java/vlaship/price/service/ParserService.java
@@ -1,0 +1,9 @@
+package vlaship.price.service;
+
+import vlaship.price.entity.Product;
+
+import java.util.List;
+
+public interface ParserService {
+    List<Product> parse(List<String> lines);
+}
