+++ src/main/java/vlaship/price/service/UpdateDbService.java
@@ -1,0 +1,5 @@
+package vlaship.price.service;
+
+public interface UpdateDbService {
+    void update();
+}
