+++ src/main/java/vlaship/price/service/UploadDataService.java
@@ -1,0 +1,9 @@
+package vlaship.price.service;
+
+import vlaship.price.entity.Product;
+
+import java.util.List;
+
+public interface UploadDataService {
+    void upload(List<Product> products);
+}
